+++ app/Api.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE InstanceSigs #-} -- for parseQueryParam :: Text -> Either Text Region
+{-# LANGUAGE InstanceSigs #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TypeOperators #-}
@@ -13,37 +13,11 @@ module Api ( API, API'
            ) where 
 
 import           Data.Aeson
-import qualified Data.ByteString.Lazy    as L
 import           Data.ByteString.Builder as B
-import qualified Data.Text               as T (Text, unpack, pack)
+import qualified Data.Text               as T (Text, unpack)
 import           Servant.API
 import           Lucid (Html)
-
-import qualified Db (Server)
-
-data EvStream
-instance Accept EvStream where
-    contentType _ = "text/event-stream"
-instance MimeRender EvStream String where
-    mimeRender _ ev = toLazyByteString $ stringUtf8 ev <> charUtf8 '\n' <> charUtf8 '\n'
-instance MimeRender EvStream ServerUpdate where -- build a JSON response for EvStream
-    mimeRender _ status = toLazyByteString $ 
-        "event: status\n" <> "data: " <> stringUtf8 (mkJson status) <> stringUtf8 "\n\n"
-
-        where   
-            mkJson (ServerUpdate reg addr port m p c q) = "{\"addr\":\"" ++ addr ++ "\",\"port\":\"" ++ port ++ "\"" ++
-                                                              apIf "m" m ++ apIf "p" p ++ apIf "c" c ++ apIf "q" q ++ "}"
-            mkJson (ServerUpdateKeepalive _)            = "{}"
-
-            apIf key (Just x) = ",\"" ++ key ++ "\":\"" ++ showIntStr x ++ "\""
-            apIf _ Nothing    = ""
-
-class ShowIntStr a where
-    showIntStr :: a -> String 
-instance ShowIntStr Int where
-    showIntStr = show
-instance ShowIntStr String where
-    showIntStr = id
+import           Db (Server(Server))
 
 -- | A partial server update packet identified by address and port, carrying optional fields
 data ServerUpdate = 
@@ -58,45 +32,39 @@ data ServerUpdate =
     , queued   :: Maybe Int    -- ^ Players in queue
     }
 
-instance ToJSON Db.Server
+class ShowIntStr a where
+    showIntStr :: a -> String 
+instance ShowIntStr Int where
+    showIntStr = show
+instance ShowIntStr String where
+    showIntStr = id
+
+data EvStream
+instance Accept EvStream where
+    contentType _ = "text/event-stream"
+instance MimeRender EvStream String where
+    mimeRender _ ev = toLazyByteString $ stringUtf8 ev <> charUtf8 '\n' <> charUtf8 '\n'
+instance MimeRender EvStream ServerUpdate where -- build a JSON response for EvStream
+    mimeRender _ status = toLazyByteString $ 
+        "event: status\n" <> "data: " <> stringUtf8 (mkJson status) <> stringUtf8 "\n\n"
+
+        where  
+            mkJson (ServerUpdate reg addr port m p c q) = "{\"addr\":\"" ++ addr ++ "\",\"port\":\"" ++ port ++ "\"" ++
+                                                                        showIf "m" m ++ showIf "p" p ++ showIf "c" c ++ showIf "q" q ++ "}"
+            mkJson (ServerUpdateKeepalive _)            = "{}"
+        
+            showIf key (Just x) = ",\"" ++ key ++ "\":\"" ++ showIntStr x ++ "\""
+            showIf _ Nothing    = ""
+
 instance FromJSON Db.Server
+instance ToJSON Db.Server where
+    toJSON (Db.Server reg name addr port m p c q t) = 
+        object  [ "addr" .= addr, "port" .= port, "name" .= name, "map" .= m, 
+                  "players" .= p, "capacity" .= c, "queued" .= q, "started" .= t
+                ]
 
-instance ToJSON ServerUpdate where
-    toJSON (ServerUpdate region addr port map players capacity queued) =
-        -- If two key-values ('addr') are clashing, the later keys and their values are preferred
-        -- Use this to prune fields that have no values from the JSON response
-        let mapPair = case map of
-                        (Just m) -> ("map", String (T.pack m))
-                        Nothing -> ("addr", String "dummy")
-            playerPair = case players of
-                        (Just p) -> ("players", String (T.pack . show $ p))
-                        Nothing -> ("addr", String "dummy") 
-            capacityPair = case capacity of
-                        (Just c) -> ("players", String (T.pack . show $ c))
-                        Nothing -> ("addr", String "dummy")
-            queuePair = case queued of
-                        (Just q) -> ("players", String (T.pack . show $ q))
-                        Nothing -> ("addr", String "dummy")
-
-        in object ([mapPair] ++ [playerPair] ++ [capacityPair] ++[queuePair] ++ ["addr" .= addr,"port" .= port])
-
-    toJSON (ServerUpdateKeepalive region) = object []
-
-{-
-instance FromJSON ServerUpdate where
-    parseJSON (Object json) = ServerUpdate 
-        <$> json .: "addr"
-        <*> json .: "port" 
--}
 data HTML
--- | Region query returns a datatype
-data Region = EU | NA | OZ | UNDEF
 
-{-| The API is made up from a static file server (CSS, JS, PNG etc), 
-    HTML server (user interface) and API server (code interface).
-    We split the API and server in two parts so that we don't have
-    to generate any JS for the static parts of the application.
--}
 type API  = ApiEndpoint
 type API' = API :<|> (Static :<|> RootEndpoint)
 
@@ -115,6 +83,8 @@ type RootEndpoint = (       Get '[HTML] String    -- static index page
                     :<|>    "dm" :> QueryParam "region" Region :> Get '[HTML] (Html ()) -- dm web app
                     )
 
+
+data Region = EU | NA | OZ | UNDEF
 
 instance FromHttpApiData Region where
     parseQueryParam :: T.Text -> Either T.Text Region