+++ app/App.hs
@@ -20,7 +20,8 @@ import Lucid (Html, source_)
 
 import Api (API, API', Region, regionToString, ServerUpdate(ServerUpdate, ServerUpdateKeepalive))
 import qualified Db (getAllByRegion, Server, getAllByRegionParsed)
-import qualified Render (header, body1, body2)
+import qualified Render (header, body)
+import qualified Data.ByteString as Render
 
 server :: FilePath -> FilePath -> Chan ServerUpdate -> Server API
 server db statics chan = status :<|> graphics
@@ -46,39 +47,22 @@ server db statics chan = status :<|> graphics
         graphics img width height = return "test"
         
 server' :: FilePath -> FilePath -> Chan ServerUpdate -> Server API'
-server' db  statics chan = server db statics chan :<|> (static :<|> root) 
-    where 
-        static = serveDirectoryWebApp statics
-        root = index :<|> dm 
-            -- | Render app server-side to memory once during startup
-            where 
-                -- | Serve the static front page
-                index :: Handler String
-                index = liftIO $ readFile $ statics ++ "/index.html"
+server' db  statics chan = server db statics chan :<|> (serveDirectoryWebApp statics :<|> (index :<|> dm )) 
+    where -- Render app server-side to memory once during startup
+        index :: Handler String -- | Serve the static front page
+        index = liftIO $ readFile $ statics ++ "/index.html"
 
-                -- TODO: render the app differently for other devices (pc vs tablet vs phone),
-                -- or maybe per the user clients request.
-                -- | Send a cacheable web app which provides an interface to query DmServers DB.
-                dm :: Maybe Region -> Handler (Html ())
-                dm Nothing    = return $ Render.header statics <> Render.body1
-                dm (Just reg) = do
-                        servers <- liftIO $ Db.getAllByRegionParsed (regionToString reg) db
-                        return $ Render.header statics <> Render.body2 servers
-
-{-
-    Idea: add a middleware layer to check the clients request header for a JSON web token. 
-    If the token exists, it means that the client has the site cached and loaded (because the token
-    is also saved in the clients offline cache with the static files), therefore the server has 
-    to send only the JSON response to the client querying the API. If the token doesn't exist, 
-    e.g., the client doesn't have the neccessary static files (.css, .html, .js etc) to operate 
-    the app, then the middleware will first push (HTTP/2) the files to the client and afterwards
-    let the application create and push a response to the API query.
--}
+        -- TODO: render the app differently for other devices (pc vs tablet vs phone),
+        -- or maybe per the user clients request.
+        -- | Send a cacheable web app which provides an interface to query DmServers DB.
+        dm :: Maybe Region -> Handler (Html ())
+        dm _ = return interface
 
 api :: Proxy API
 api = Proxy
 api' :: Proxy API'
 api' = Proxy
 
-app :: (FilePath, FilePath, Chan ServerUpdate) -> Application
 app (statics, db, chan) = serve api' (server' db statics chan)
+
+interface = Render.header <> Render.body